+++ App/Entities/Swagger/2.0/SwaggerExternalDocs.cs
@@ -1,0 +1,11 @@
+namespace App.Entities.Swagger.Two
+{
+    public class SwaggerExternalDocs : ISwaggerProperty
+    {
+        public bool IsValid => !string.IsNullOrEmpty(Url);  
+
+        // Structure
+        public string Description { get; set; }
+        public string Url { get; set; }
+    }
+}
