+++ App/Entities/Swagger/2.0/SwaggerHeader.cs
@@ -1,0 +1,7 @@
+namespace App.Entities.Swagger.Two
+{
+    public class SwaggerHeader : SwaggerItems, ISwaggerProperty
+    {
+        public string Description { get; set; }
+    }
+}
