+++ App/Entities/Swagger/Generic/IDocumentation.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace App.Entities.Swagger
+{
+    public interface IDocumentation : ISwaggerProperty
+    {
+        Dictionary<string, ISwaggerPathItem> GetPaths();
+    }
+}
