+++ App/Entities/Swagger/Generic/ISwaggerOperation.cs
@@ -1,0 +1,4 @@
+namespace App.Entities.Swagger
+{
+    public interface ISwaggerOperation : ISwaggerProperty { }
+}
