+++ App/Entities/Swagger/Generic/ISwaggerProperty.cs
@@ -1,0 +1,7 @@
+namespace App.Entities.Swagger
+{
+    public interface ISwaggerProperty
+    {
+        bool IsValid { get; }
+    }
+}
