--- App/Entities/Swagger/IDocumentation.cs
@@ -1,7 +1,0 @@
-namespace App.Entities.Swagger
-{
-    public interface IDocumentation
-    {
-
-    }
-}
