+++ App/Services/Validations/Level1/ValidateGetRoutesPerPath.cs
@@ -6,17 +6,17 @@ namespace App.Services.Validations.Level1
 {
     public class ValidateGetRoutesPerPath : IValidateGetRoutesPerPath
     {     
-        public ValidationOutput Validate(Documentation documentation)
+        public ValidationOutput Validate(IDocumentation documentation)
         {
             var output = new ValidationOutput();
 
-            var groupedByFirstWord = documentation.Paths
+            var groupedByFirstWord = documentation.GetPaths()
                 .GroupBy(path => new ApiPath(path.Key).Parts[0].ToString())
-                .OrderBy(paths => paths.Count(path => path.Value.Get != null));
+                .OrderBy(paths => paths.Count(path => path.Value.GetOperations()[HTTPVERBS.GET] != null));
 
             foreach (var group in groupedByFirstWord)
             {
-                int getsQuantity = group.Count(path => path.Value.Get != null);  
+                int getsQuantity = group.Count(path => path.Value.GetOperations()[HTTPVERBS.GET] != null);  
 
                 if (getsQuantity > 2)
                     output.AddProblem($"{group.Key} has {getsQuantity} GET routes, the maximum is 2.");
