+++ App/Services/Validations/Level1/ValidateIdentifiers.cs
@@ -8,12 +8,12 @@ namespace App.Services.Validations.Level1
 {
     public class ValidateIdentifiers : IValidateIdentifiers
     {
-        public ValidationOutput Validate(Documentation documentation)
+        public ValidationOutput Validate(IDocumentation documentation)
         {
             var output = new ValidationOutput();
 
             //Getting paths from document
-            var rawPaths = documentation.Paths.Keys;
+            var rawPaths = documentation.GetPaths().Keys;
             var paths = rawPaths.Select(s => new ApiPath(s));
 
             //Validating
