+++ App/Services/Validations/Level1/ValidatePathLevels.cs
@@ -13,12 +13,12 @@ namespace App.Services.Validations.Level1
         //Default = 2
         public ValidatePathLevels(int limit) { this.limit = limit; }
 
-        public ValidationOutput Validate(Documentation documentation)
+        public ValidationOutput Validate(IDocumentation documentation)
         {
             var output = new ValidationOutput();
 
             //Getting paths from document
-            var rawPaths = documentation.Paths.Keys;
+            var rawPaths = documentation.GetPaths().Keys;
             var paths = rawPaths.Select(s => new ApiPath(s));
 
             //Validating
