+++ App/Services/Validations/Level3/ValidatePathHttpVerbs.cs
@@ -14,13 +14,13 @@ namespace App.Services.Validations.Level3
             _allowedVerbs = allowedVerbs;
         }
 
-        public ValidationOutput Validate(Documentation documentation)
+        public ValidationOutput Validate(IDocumentation documentation)
         {
             var output = new ValidationOutput();
 
-            foreach (var path in documentation.Paths)
+            foreach (var path in documentation.GetPaths())
             {
-                var verbsNotAllowed = path.Value.Verbs.Except(_allowedVerbs);
+                var verbsNotAllowed = path.Value.GetVerbs().Except(_allowedVerbs);
                 if (verbsNotAllowed.Count() > 0)
                     output.AddProblem($"Path {path.Key} has verbs that are not allowed: {string.Concat(verbsNotAllowed.Select(verb => verb + " "))}");
             }
