+++ Tests/Entities/Swagger/DocumentationIsValid.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using App.Entities.Swagger;
+using App.Entities.Swagger.Two;
 using Xunit;
 
 namespace Tests
