+++ Tests/Services/Validations/ValidateIdentifier.cs
@@ -1,7 +1,7 @@
 using Xunit;
 using System.Linq;
 using System.Collections.Generic;
-using App.Entities.Swagger;
+using App.Entities.Swagger.Two;
 using App.Services.Validations.Level1;
 
 namespace Tests
